pyproject.toml
[project]
name = "zerodha-mcp"
version = "0.1.0"
description = "MCP server for Zerodha account - query funds, holdings, positions, orders, and P&L"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "requests>=2.28.0",
    "pyotp>=2.8.0",
    "python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
zerodha-mcp = "server:mcp.run"
